Verificando campos input post, permissões de acesso a paginas

+++ myblog/post/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect
-from accounts.models import Post, Categoria, Usuario, UpdatePost
+from accounts.models import Post, Categoria, Usuario
 from accounts.views import login
+from django.core.paginator import Paginator
 
 
 # Create your views here.
@@ -9,7 +10,13 @@ def post(request):
     usuario_name = request.session['usuario_name']
     usuario_tipo = request.session['usuario_tipo']
     
-    post = Post.objects.all()
+    post = Post.objects.order_by('-date_create')
+    
+    paginator = Paginator(post,10)
+
+    page_number = request.GET.get('page')
+    post = paginator.get_page(page_number)
+    
     return render(request, 'post/post.html',{
         'posts': post,
         'usuario_id':usuario_id,
@@ -19,6 +26,7 @@ def post(request):
 
 def view_post(request,post_id):
     post = Post.objects.get(id=post_id)
+    
     return render(request, 'post/view_post.html',{
         'post': post
     })
@@ -33,12 +41,15 @@ def post_create(request):
     categoria_name = request.POST.get('categoria')
     titulo = request.POST.get('titulo')
     publicacao = request.POST.get('publicacao')
-
+    
     if usuario_tipo != 'AU':
         return redirect('post')
     
     try:
         if request.method == 'POST':
+            if not categoria_name or not titulo or not publicacao:
+                return redirect('post_create')
+
             categoria = Categoria.objects.filter(nome_categoria=categoria_name).first()
             autor = Usuario.objects.filter(id=usuario_id).first()
             post = Post(categoria=categoria,autor=autor,titulo=titulo,publicacao=publicacao)
@@ -55,6 +66,11 @@ def post_create(request):
 def post_update(request,post_id):
     id_post = Post.objects.filter(id=post_id).first()
     categorias = Categoria.objects.all()
+
+    usuario_tipo = request.session['usuario_tipo']
+
+    if usuario_tipo != 'AU':
+        return redirect('post')
     
     if request.method == 'POST':
         try:
@@ -76,6 +92,11 @@ def post_update(request,post_id):
     })
 
 def post_delete(request,post_id):
+    usuario_tipo = request.session['usuario_tipo']
+
+    if usuario_tipo != 'AU':
+        return redirect('post')
+
     id_post = Post.objects.get(id=post_id)
     id_post.delete()
     print('Deletou um usuario com sucesso!')
